Move auth and boilerplate into helper file
